% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{batch_detect_dominant_language}
\alias{batch_detect_dominant_language}
\title{Determines the dominant language of the input text for a batch of documents}
\usage{
batch_detect_dominant_language(TextList)
}
\arguments{
\item{TextList}{[required] A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document should contain at least 20 characters and must contain fewer than 5,000 bytes of UTF-8 encoded characters.}
}
\description{
Determines the dominant language of the input text for a batch of documents. For a list of languages that Amazon Comprehend can detect, see \href{http://docs.aws.amazon.com/comprehend/latest/dg/how-languages.html}{Amazon Comprehend Supported Languages}.
}
\section{Accepted Parameters}{
\preformatted{batch_detect_dominant_language(
  TextList = list(
    "string"
  )
)
}
}

