% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{batch_detect_entities}
\alias{batch_detect_entities}
\title{Inspects the text of a batch of documents for named entities and returns information about them}
\usage{
batch_detect_entities(TextList, LanguageCode)
}
\arguments{
\item{TextList}{A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document must contain fewer than 5,000 bytes of UTF-8 encoded characters.}

\item{LanguageCode}{The language of the input documents. You can specify English ("en") or Spanish ("es"). All documents must be in the same language.}
}
\description{
Inspects the text of a batch of documents for named entities and returns information about them. For more information about named entities, see how-entities
}
\examples{

}
