% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{batch_detect_key_phrases}
\alias{batch_detect_key_phrases}
\title{Detects the key noun phrases found in a batch of documents}
\usage{
batch_detect_key_phrases(TextList, LanguageCode)
}
\arguments{
\item{TextList}{[required] A list containing the text of the input documents. The list can contain a maximum of 25 documents. Each document must contain fewer that 5,000 bytes of UTF-8 encoded characters.}

\item{LanguageCode}{[required] The language of the input documents. You can specify English ("en") or Spanish ("es"). All documents must be in the same language.}
}
\description{
Detects the key noun phrases found in a batch of documents.
}
\section{Accepted Parameters}{
\preformatted{batch_detect_key_phrases(
  TextList = list(
    "string"
  ),
  LanguageCode = "en"|"es"|"fr"|"de"|"it"|"pt"
)
}
}

