% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{create_document_classifier}
\alias{create_document_classifier}
\title{Creates a new document classifier that you can use to categorize documents}
\usage{
create_document_classifier(
  DocumentClassifierName = "string",
  DataAccessRoleArn = "string",
  InputDataConfig = list(
    S3Uri = "string"
  ),
  ClientRequestToken = "string",
  LanguageCode = "en"|"es"|"fr"|"de"|"it"|"pt"
)
}
\arguments{
\item{DocumentClassifierName}{[required] The name of the document classifier.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.}

\item{InputDataConfig}{[required] Specifies the format and location of the input data for the job.}

\item{ClientRequestToken}{A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.}

\item{LanguageCode}{[required] The language of the input documents. You can specify English ("en") or Spanish ("es"). All documents must be in the same language.}
}
\description{
Creates a new document classifier that you can use to categorize documents. To create a classifier you provide a set of training documents that labeled with the categories that you want to use. After the classifier is trained you can use it to categorize a set of labeled documents into the categories. For more information, see how-document-classification.
}
\examples{

}
