% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{create_entity_recognizer}
\alias{create_entity_recognizer}
\title{Creates an entity recognizer using submitted files}
\usage{
create_entity_recognizer(
  RecognizerName = "string",
  DataAccessRoleArn = "string",
  InputDataConfig = list(
    EntityTypes = list(
      list(
        Type = "string"
      )
    ),
    Documents = list(
      S3Uri = "string"
    ),
    Annotations = list(
      S3Uri = "string"
    ),
    EntityList = list(
      S3Uri = "string"
    )
  ),
  ClientRequestToken = "string",
  LanguageCode = "en"|"es"|"fr"|"de"|"it"|"pt"
)
}
\arguments{
\item{RecognizerName}{[required] The name given to the newly created recognizer. Recognizer names can be a maximum of 256 characters. Alphanumeric characters, hyphens (-) and underscores (\_) are allowed. The name must be unique in the account/region.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name (ARN) of the AWS Identity and Management (IAM) role that grants Amazon Comprehend read access to your input data.}

\item{InputDataConfig}{[required] Specifies the format and location of the input data. The S3 bucket containing the input data must be located in the same region as the entity recognizer being created.}

\item{ClientRequestToken}{A unique identifier for the request. If you don't set the client request token, Amazon Comprehend generates one.}

\item{LanguageCode}{[required] The language of the input documents. All documents must be in the same language. Only English ("en") is currently supported.}
}
\description{
Creates an entity recognizer using submitted files. After your \code{CreateEntityRecognizer} request is submitted, you can check job status using the API.
}
\examples{

}
