% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{describe_document_classification_job}
\alias{describe_document_classification_job}
\title{Gets the properties associated with a document classification job}
\usage{
describe_document_classification_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier that Amazon Comprehend generated for the job. The operation returns this identifier in its response.}
}
\description{
Gets the properties associated with a document classification job. Use this operation to get the status of a classification job.
}
\section{Accepted Parameters}{
\preformatted{describe_document_classification_job(
  JobId = "string"
)
}
}

