% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{describe_entities_detection_job}
\alias{describe_entities_detection_job}
\title{Gets the properties associated with an entities detection job}
\usage{
describe_entities_detection_job(JobId)
}
\arguments{
\item{JobId}{The identifier that Amazon Comprehend generated for the job. The operation returns this identifier in its response.}
}
\description{
Gets the properties associated with an entities detection job. Use this operation to get the status of a detection job.
}
\examples{

}
