% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{describe_entity_recognizer}
\alias{describe_entity_recognizer}
\title{Provides details about an entity recognizer including status, S3 buckets containing training data, recognizer metadata, metrics, and so on}
\usage{
describe_entity_recognizer(EntityRecognizerArn)
}
\arguments{
\item{EntityRecognizerArn}{[required] The Amazon Resource Name (ARN) that identifies the entity recognizer.}
}
\description{
Provides details about an entity recognizer including status, S3 buckets containing training data, recognizer metadata, metrics, and so on.
}
\section{Accepted Parameters}{
\preformatted{describe_entity_recognizer(
  EntityRecognizerArn = "string"
)
}
}

