% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{describe_topics_detection_job}
\alias{describe_topics_detection_job}
\title{Gets the properties associated with a topic detection job}
\usage{
describe_topics_detection_job(JobId)
}
\arguments{
\item{JobId}{[required] The identifier assigned by the user to the detection job.}
}
\description{
Gets the properties associated with a topic detection job. Use this operation to get the status of a detection job.
}
\section{Accepted Parameters}{
\preformatted{describe_topics_detection_job(
  JobId = "string"
)
}
}

