% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{detect_dominant_language}
\alias{detect_dominant_language}
\title{Determines the dominant language of the input text}
\usage{
detect_dominant_language(
  Text = "string"
)
}
\arguments{
\item{Text}{[required] A UTF-8 text string. Each string should contain at least 20 characters and must contain fewer that 5,000 bytes of UTF-8 encoded characters.}
}
\description{
Determines the dominant language of the input text. For a list of languages that Amazon Comprehend can detect, see \href{http://docs.aws.amazon.com/comprehend/latest/dg/how-languages.html}{Amazon Comprehend Supported Languages}.
}
\examples{

}
