% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{detect_entities}
\alias{detect_entities}
\title{Inspects text for named entities, and returns information about them}
\usage{
detect_entities(Text, LanguageCode)
}
\arguments{
\item{Text}{A UTF-8 text string. Each string must contain fewer that 5,000 bytes of UTF-8 encoded characters.}

\item{LanguageCode}{The language of the input documents. You can specify English ("en") or Spanish ("es"). All documents must be in the same language.}
}
\description{
Inspects text for named entities, and returns information about them. For more information, about named entities, see how-entities.
}
\examples{

}
