% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{list_dominant_language_detection_jobs}
\alias{list_dominant_language_detection_jobs}
\title{Gets a list of the dominant language detection jobs that you have submitted}
\usage{
list_dominant_language_detection_jobs(
  Filter = list(
    JobName = "string",
    JobStatus = "SUBMITTED"|"IN_PROGRESS"|"COMPLETED"|"FAILED"|"STOP_REQUESTED"|"STOPPED",
    SubmitTimeBefore = as.POSIXct("2015-01-01"),
    SubmitTimeAfter = as.POSIXct("2015-01-01")
  ),
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{Filter}{Filters that jobs that are returned. You can filter jobs on their name, status, or the date and time that they were submitted. You can only set one filter at a time.}

\item{NextToken}{Identifies the next page of results to return.}

\item{MaxResults}{The maximum number of results to return in each page. The default is 100.}
}
\description{
Gets a list of the dominant language detection jobs that you have submitted.
}
\examples{

}
