% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{list_key_phrases_detection_jobs}
\alias{list_key_phrases_detection_jobs}
\title{Get a list of key phrase detection jobs that you have submitted}
\usage{
list_key_phrases_detection_jobs(Filter = NULL, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{Filter}{Filters the jobs that are returned. You can filter jobs on their name, status, or the date and time that they were submitted. You can only set one filter at a time.}

\item{NextToken}{Identifies the next page of results to return.}

\item{MaxResults}{The maximum number of results to return in each page. The default is 100.}
}
\description{
Get a list of key phrase detection jobs that you have submitted.
}
\examples{

}
