% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_package.R
\docType{package}
\name{paws.comprehend-package}
\alias{paws.comprehend}
\alias{paws.comprehend-package}
\title{paws.comprehend: Amazon Comprehend}
\description{
Amazon Comprehend is an AWS service for gaining insight into
the content of documents. Use these actions to determine the topics
contained in your documents, the topics they discuss, the predominant
sentiment expressed in them, the predominant language used, and more.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
