% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehend_operations.R
\name{start_dominant_language_detection_job}
\alias{start_dominant_language_detection_job}
\title{Starts an asynchronous dominant language detection job for a collection of documents}
\usage{
start_dominant_language_detection_job(
  InputDataConfig = list(
    S3Uri = "string",
    InputFormat = "ONE_DOC_PER_FILE"|"ONE_DOC_PER_LINE"
  ),
  OutputDataConfig = list(
    S3Uri = "string"
  ),
  DataAccessRoleArn = "string",
  JobName = "string",
  ClientRequestToken = "string"
)
}
\arguments{
\item{InputDataConfig}{[required] Specifies the format and location of the input data for the job.}

\item{OutputDataConfig}{[required] Specifies where to send the output files.}

\item{DataAccessRoleArn}{[required] The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that grants Amazon Comprehend read access to your input data. For more information, see \url{https://docs.aws.amazon.com/comprehend/latest/dg/access-control-managing-permissions.html#auth-role-permissions}.}

\item{JobName}{An identifier for the job.}

\item{ClientRequestToken}{A unique identifier for the request. If you do not set the client request token, Amazon Comprehend generates one.}
}
\description{
Starts an asynchronous dominant language detection job for a collection of documents. Use the operation to track the status of a job.
}
\examples{

}
